+++ src/AngleSharp.Performance.Utilities/Properties/AssemblyInfo.cs
@@ -1,8 +1,0 @@
-using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyCopyright("Copyright © AngleSharp, 2013-2026")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: ComVisible(false)]
-[assembly: Guid("aa7bce2a-f953-4d04-8bed-69eb73687b22")]